Added new UI for history exams list view.
The new history view adds the ability to review, retake
and resume directly from the history of exams.
Added review, retake and resume buttons in the history list.
If there is only one attempt, then review will directly show
the detailed review of the attempt.
Retake button can be used to directly restart the exam.
Resume button in the list view shows list of paused attempts to
resume.