Fix end exam degrade
End exam was not working properly after migrating to 2.2 APIs.
Problem was the API was not returning all the fields of the exam
in the attempts model. The attempts_count and paused_attempts_count
were not part of the exam field in the attempts model. As a result,
parceleable was failing in trying to convert null to int for these
fields during starting ReviewActivity.
Now the API is changed not to include exam json as part of attempts
model. Instead exam_url is returned.
The exam field in the Attempt model is now removed to reflect the
same.