Remove retake option from attempts list and review
The retake icon in the toolbar was confusing in terms of
usability and also not honouring the allow_retake option from
the exam. This is now removed.
User can no longer retake from the attempts list and review
toolbar icon. This functionality is moved to history of exams
list item as a button.