Add support for email MCQs in available exams list
Available exams now has option to email MCQs PDF if option
is enabled in the exam. The UI is now similar to that of
history list. Upcoming and Available exams now have separate
adapters to allow more customization.
The exam model adds support for whether to allow pdf download
or not.