Add email pdf button in ReviewStatsFragment
Previously email pdf button was only available in the
attempts list view. As a result of this, there was no
interface to email the pdf when an exam was attempted
only once. Now the email pdf action can also be triggered
from the ReviewStatsFragment. It also honours allow_pdf
setting of the exam.