Fix crash when number of posts is < 3
If the number of posts is < 3 the app was crashing. The problem
seem to come from onScrollListener callback being called before loader and
adapter are initialized first. The getLastVisiblePosition code returs
-1 which satisfies the if condtion in onScrollListener and tries
to restart the loader which wasn't initialized at the first place.

This is now fixed by calling initLoader before enabling
onScrollListener.