Fix password hint typeface changing to monospace for edittext
Due to a bug in android, typeface of edittext with type password
changes to monospace.
For now we have used fontFamily to fix this issue. However, this
still doesn't work in sdk level < 16. The solutions given to mitigate
this issue in the internet doesn't seem to work for AppCompatEditText.
So for now living with this known limitation.
http://stackoverflow.com/questions/24117178/android-typeface-is-changed-when-i-apply-password-type-on-edittext
http://stackoverflow.com/questions/3406534/password-hint-font-in-android?rq=1