Fix questions not displaying for IBPS template exam
If the total number of subjects in an exam is < 2, then
questions were not displayed in the exam during an attempt.
The reason was the questions were not added to the local
attemptsItemList if number of subjects are < 2. This is now
fixed to only hide the subjects filter if number of subjects
< 2.