Cleanup product details UI with support for external payment.
Product details page is now cleaned up to be much more UX
friendly.
RawExam model is used as the exams part of the products will
not contain attempt related fields.
The adapters for Exams list and Notes list are modified to give
a much more aesthetic appearance.
If external payment link is given, then the app delegates the job
by calling an external browser.