Hide product(store) & fix the bug in posts
Hide the store icon in home screen.
If user went to the post activity by clicking the category from quick
link, & if there is no posts in that category in the db, then that
category name is not selecting in the spinner. Now it is fixed.

Loading progress bar appear in the bottom of the list view while scroll
is also showing if no items in list also. Now it is fixed.