Load post categories using pagination and Fix authorization bug
Sometimes pager which loads older posts is initialized with
TestpressService, before refreshPager initializes the TestpressService
with auth-token. Therefore auth-token is missing in the API calls which
load the older posts.
This is fixed by getting the TestpressService in all the occurrences
using the getTestpressService method which initializes the
TestpressService with auth-token for the first time & returns the same
instance afterwards.