Force user data in android. (#99)
### Changes done
Added checking condition for force data if the user is authenticated. Every time the user will interact with MainActivity. Codes will check for the institute force setting. If it's true it will ping the API to check is data is collected for the user. If not so it will hide all the UI and will ping sso_url and open the web view to fill the values.

If the user tries to press the back button, it does all the above process in onResume function. Will again redirect to web view if it didn't find user has filed all the values.

If some error occurs during pinging the CHECK_PERMISSION_URL or SSO URL. It will hide all the UI and will ask to again refresh the page.

### Reason for the changes
We have added the feature force data in the web. But still, we were not for